+++ src/Sulu/Bundle/ContentBundle/Tests/Functional/Controller/NodeResourcelocatorControllerTest.php
@@ -55,6 +55,7 @@ class NodeResourcelocatorControllerTest extends SuluTestCase
         $data = [
             [
                 'title' => 'Produkte',
+                'template' => 'default',
                 'tags' => [
                     'tag1',
                     'tag2',
@@ -64,6 +65,7 @@ class NodeResourcelocatorControllerTest extends SuluTestCase
             ],
             [
                 'title' => 'News',
+                'template' => 'default',
                 'tags' => [
                     'tag1',
                     'tag2',
@@ -73,6 +75,7 @@ class NodeResourcelocatorControllerTest extends SuluTestCase
             ],
             [
                 'title' => 'test',
+                'template' => 'default',
                 'tags' => [
                     'tag1',
                     'tag2',
@@ -82,6 +85,7 @@ class NodeResourcelocatorControllerTest extends SuluTestCase
             ],
             [
                 'title' => 'test-2',
+                'template' => 'default',
                 'tags' => [
                     'tag1',
                     'tag2',
@@ -91,6 +95,7 @@ class NodeResourcelocatorControllerTest extends SuluTestCase
             ],
             [
                 'title' => 'test',
+                'template' => 'default',
                 'tags' => [
                     'tag1',
                     'tag2',
@@ -107,15 +112,15 @@ class NodeResourcelocatorControllerTest extends SuluTestCase
                 'PHP_AUTH_PW' => 'test',
             ]
         );
-        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en&template=default', $data[0]);
+        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en', $data[0]);
         $data[0] = (array) json_decode($client->getResponse()->getContent(), true);
-        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en&template=default', $data[1]);
+        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en', $data[1]);
         $data[1] = (array) json_decode($client->getResponse()->getContent(), true);
-        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en&template=default&parent=' . $data[1]['id'], $data[2]);
+        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en&parent=' . $data[1]['id'], $data[2]);
         $data[2] = (array) json_decode($client->getResponse()->getContent(), true);
-        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en&template=default&parent=' . $data[1]['id'], $data[3]);
+        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en&parent=' . $data[1]['id'], $data[3]);
         $data[3] = (array) json_decode($client->getResponse()->getContent(), true);
-        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en&template=default&parent=' . $data[3]['id'], $data[4]);
+        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en&parent=' . $data[3]['id'], $data[4]);
         $data[4] = (array) json_decode($client->getResponse()->getContent(), true);
 
         return $data;
@@ -167,7 +172,7 @@ class NodeResourcelocatorControllerTest extends SuluTestCase
         $newsData['url'] = '/test';
         $this->client->request(
             'PUT',
-            '/api/nodes/' . $newsData['id'] . '?webspace=sulu_io&language=en&template=default',
+            '/api/nodes/' . $newsData['id'] . '?webspace=sulu_io&language=en',
             $newsData
         );
         $newsData = (array) json_decode($this->client->getResponse()->getContent(), true);
@@ -191,7 +196,7 @@ class NodeResourcelocatorControllerTest extends SuluTestCase
         $newsData['url'] = '/test';
         $this->client->request(
             'PUT',
-            '/api/nodes/' . $newsData['id'] . '?webspace=sulu_io&language=en&template=default',
+            '/api/nodes/' . $newsData['id'] . '?webspace=sulu_io&language=en',
             $newsData
         );
         $newsData = (array) json_decode($this->client->getResponse()->getContent());
@@ -227,7 +232,7 @@ class NodeResourcelocatorControllerTest extends SuluTestCase
         $newsData['url'] = '/test';
         $this->client->request(
             'PUT',
-            '/api/nodes/' . $newsData['id'] . '?webspace=sulu_io&language=en&template=default',
+            '/api/nodes/' . $newsData['id'] . '?webspace=sulu_io&language=en',
             $newsData
         );
         $newsData = (array) json_decode($this->client->getResponse()->getContent());